Drop the Generic prefix
